Rename module to remote_config
